requirements.txt
google-adk
requests
fastapi        # or FastAPI for web_demo
python-dotenv
